refactor(notes): extract use_document_import hook from NoteDetail
